When store new file, if we have thumbnail, we change filename to thumbnail for client side use
if we have thumbnail, we change filename to thumbnail for client side to display. This is the same behavior as index action in this controller
Please consider whether or not merge this.
I use vuejs as the front end data-driven app dev. When uploaded file, if this $file with thumb returned, client side will update view automatically with this returned data. Otherwise, no thumb image can be retrieved